Pull validation logic into the messages
